Uploading NIDS and system default causing noise
Let's have both as as static value as having this as a random value could lower the accuracy of the model too much. Let's have it this way for at least default
Currently,  NIDS' success rate is 80%, so 20% of the time, it will not detect the attacks and will not run the take action function in the envRL.py.